backend/requirements.txt
fastapi
uvicorn
pydantic
google-genai
opencv-python
pyaudio
pillow
mss
websockets
python-dotenv
sarvamai
